Start and finalize delivery implemented

+++ src/app/controllers/DeliverymanController.js
@@ -44,6 +44,14 @@ class DeliverymanController {
   }
 
   async show(req, res) {
+    const schema = Yup.object().shape({
+      id: Yup.number().positive(),
+    });
+
+    if (!(await schema.isValid({ ...req.params }))) {
+      return res.status(400).json({ error: 'Data validation fails' });
+    }
+
     const deliveryman = await Deliveryman.findByPk(req.params.id, {
       attributes: ['id', 'name', 'email'],
       include: [
@@ -65,10 +73,11 @@ class DeliverymanController {
     const schema = Yup.object().shape({
       name: Yup.string(),
       email: Yup.string(),
-      avatar_id: Yup.number(),
+      avatar_id: Yup.number().positive(),
+      id: Yup.number().positive(),
     });
 
-    if (!(await schema.isValid(req.body))) {
+    if (!(await schema.isValid({ ...req.body, ...req.params }))) {
       return res.status(400).json({ error: 'Data validation fails' });
     }
 
@@ -94,6 +103,14 @@ class DeliverymanController {
   }
 
   async delete(req, res) {
+    const schema = Yup.object().shape({
+      id: Yup.number().positive(),
+    });
+
+    if (!(await schema.isValid({ ...req.params }))) {
+      return res.status(400).json({ error: 'Data validation fails' });
+    }
+
     const deliveryman = await Deliveryman.destroy({
       where: {
         id: req.params.id,
